refactor: split oversized economy test support modules (Refs #4108)
Slice B de-parts extraction fixtures (tile vs game), resource-extractor
scenario runner vs tables, and DealMatcher inputs vs row builders while
keeping back-compat barrels and shaving support LOC to 7435.
